+++ lib/custom_refresh_indicator.dart
@@ -2,4 +2,5 @@ export 'src/data/data.dart';
 export 'src/custom_refresh_indicator.dart';
 export 'src/delegates/delegates.dart';
 export 'src/widgets/widgets.dart';
+export 'src/physics/physics.dart';
 export 'src/utils/utils.dart';